requirements.txt
# === Core OpenAI & LangChain Stack ===
openai==2.7.1
langchain==1.0.4
langgraph==1.0.2
langchain-openai==1.0.2
langchain-community==0.4.1
langchain-text-splitters==1.0.0

# === Vector Store (FAISS) ===
faiss-cpu==1.12.0

# === Web Framework & Server ===
fastapi==0.121.0
uvicorn[standard]==0.38.0

# === UI (Gradio) ===
gradio==5.49.1

# === PDF & 문서 처리 ===
pypdf==6.1.3

# === Tokenization ===
tiktoken==0.12.0

# === 기타 유틸 ===
python-dotenv==1.0.1
